+++ src/cqspoolterrain/__init__.py
@@ -1,12 +1,15 @@
 from .industrialStairs import industrial_stairs
 from .Spool import Spool
 from .Cradle import Cradle
+
 from .Stairs import Stairs
 from .StairLift import StairLift
+
+from .SteelFrame import SteelFrame
 from .ControlPlatform import ControlPlatform
+
 from .SpoolCladding import SpoolCladding
 from .SpoolCladdingGreebled import SpoolCladdingGreebled
 from .SpoolCladdingGreebledUnique import SpoolCladdingGreebledUnique
-
 
 from .PowerStation import PowerStation